requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
transformers
faiss-cpu
sentence-transformers
accelerate
torch 
torchvision 
torchaudio 
--index-url https://download.pytorch.org/whl/cu118
